fix: add schema folder and lib folder, moved schemas into schema.js and moved login/registration handlers into api.js
