+++ src/Components/SignIn/SignIn.jsx
@@ -1,32 +1,21 @@
 import React from 'react'
 import Input from '../Input/Input'
 import { useForm } from 'react-hook-form'
-import z from "zod"
 import { zodResolver } from '@hookform/resolvers/zod'
+import { singinSchema } from '../../Schemas/Schemas';
+import { handleSignIn } from '../../lib/Api/Api';
 
-const singinSchema = z.object({
-    email: z.string().email().min(1),
-    password: z.string().min(8)
-})
 
 function SignIn() {
     const { register, handleSubmit, formState: { errors, isLoading } } = useForm({
         resolver: zodResolver(singinSchema)
     });
-    const handleSignIn = (data)=>{
-        const email = localStorage.getItem('email');
-        const password = localStorage.getItem("password")
-        if(email === data.email && password ===data.password){
-            console.log("succesfully authorized!");
-        }else{
-            console.log("wrong credentials!");
-        }
-    }
+
     return (
         <form onSubmit={handleSubmit(handleSignIn)} className="flex flex-col gap-5">
             <div className="signup flex flex-col gap-y-5">
                 <div>
-                    <Input {...register("email")} id="email" type="email" label="Input Email"  placeholder="Enter your email." />
+                    <Input {...register("email")} id="email" type="email" label="Input Email" placeholder="Enter your email." />
                     {errors.email?.message && <p className='text-red-600'>{errors.email.message}</p>}
 
                 </div>
